improvements

+++ par2/par2-recreate-for-outdated.sh
@@ -5,6 +5,7 @@ scriptdir="$(dirname "$0")"
 scriptdir="$( realpath "$scriptdir" || readlink -f "$scriptdir" || { cd "$scriptdir" >/dev/null 2>&1 || exit 2 ; pwd || exit 2 ; cd - >/dev/null 2>&1  || exit 2 ; } )"
 
 echo "This is a dangerous operation. Wait 30 seconds if you really want to do this, or ctrl+c NOW to stop it."
+echo "(Dangerous because you will lose the PAR2 files that can recover your data. First check that there is nothing you want to recover.)"
 sleep 30
 
 
@@ -13,13 +14,15 @@ sleep 30
 
 ## find files and folders that are newer than their respective PAR2 files and delete the PAR2 files there
 
-find . -type f -name ".PAR2PROTECT_DPT1" -exec bash -c '
+find . -type f -name ".PAR2PROTECT_DPT1" -not -path '*/EXAMPLE-EXCLUDE-PATH/Files/backups-tmp-home/*' -exec bash -c '
     cd "$( dirname "$1" )" && 
     find . -mindepth 1 -maxdepth 1 -type d -exec bash -c '"'"'
         cd "$1" && 
-        if [[ "$( find "$(pwd)" -newer PAR2.par2 )" != "" ]] ; then
-            pwd && 
-            rm -f PAR2.par2 PAR2.vol*.par2 ; 
+        if [[ -e "PAR2.par2" ]] ; then
+            if [[ "$( find "$(pwd)" -newer PAR2.par2 )" != "" ]] ; then
+                pwd && 
+                rm -f PAR2.par2 PAR2.vol*.par2 ; 
+            fi ;
         fi ;
     '"'"' _ {""} ";" 
 ' _ {} ";" 
@@ -30,16 +33,18 @@ find . -type f -name ".PAR2PROTECT_DPT1" -exec bash -c '
 
 ## find files and folders that are newer than their respective PAR2 files and delete the PAR2 files there
 
-find . -type f -name ".PAR2PROTECT_NOR" -exec bash -c '
+find . -type f -name ".PAR2PROTECT_NOR" -not -path '*/EXAMPLE-EXCLUDE-PATH/Files/backups-tmp-home/*' -exec bash -c '
     cd "$( dirname "$1" )" && 
-    if [[ "$( 
-        { 
-            find "$(pwd)" -maxdepth 0  -type d -newer PAR2.par2 ; 
-            find "$(pwd)" -maxdepth 1 -type f -newer PAR2.par2 ; 
-        } 
-    )" != "" ]] ; then
-        pwd && 
-        rm -f PAR2.par2 PAR2.vol*.par2 ; 
+    if [[ -e "PAR2.par2" ]] ; then
+        if [[ "$( 
+            { 
+                find "$(pwd)" -maxdepth 0  -type d -newer PAR2.par2 ; 
+                find "$(pwd)" -maxdepth 1 -type f -newer PAR2.par2 ; 
+            } 
+        )" != "" ]] ; then
+            pwd && 
+            rm -f PAR2.par2 PAR2.vol*.par2 ; 
+        fi ;
     fi ;
 ' _ {} ";"
